sessions: modify datatype tests cleanup path
With the changes for sessions, OMPI uses a different way
of cleaning up allocated resources,namely using a
registration of cleanup callbacks that are invoked
when opal_finalize_cleanup_domain is called on the
list of callbacks.  This obviates the need, for example,
for a test to explicitly cleanup datatype related resources
by invoking ompi_datatype_finalize explicitly.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>