c-bindings: add support for const request args
Three new functions were added to the MPI API as part of the 4.1 standard.
These used an array of const MPI_Request s which the bindings code
didn't support.

This commit also fixes back the mpi.h prototypes to include the constants
and required changes to the new template files.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>