ABI: Switch to using MPI Standard ABI values
This mod switches from using "synthetic" defined values and handles for
the ones specified in the MPI 5.1 standard.

The python infrastructure included in this PR for generating both a
"canonical" abi standard compatible MPI 5.1 mpi.h using two json files:

1) mpi-standard-abi.json
2) mpi-standard-apis.json

2 is generated as part of building the MPI standard.  We import that into our project
for use in generating both the mpi.h as well as interface definitions in the man pages.

1 is generated using a separate script that processes the tables in Appendix A of
standard.  Ideally this script will be merged into the MPI standard code base at some
point.  This script is currently at https://github.com/Joe-Downs/mpi-standard/tree/pr/handle-constant-tool/const-tool .
It is used to generate the portion of mpi.h where defined values and handles are specified.

The converter functions that had been generated as part of the build out of the abi variants
of the 'c' MPI interfaces are no define in persistent file.  The methods
defined in this file will be optimized to make use of the Huffman code characteristics of
the predefined values in a subsequent PR.

This commit also enables generation of the abi interfaces and header files by default.

Signed-off-by: Joseph Downs <joe.downs@lanl.gov>